broadcast-monitoring/src/audio_detect/app/requirements.txt
-i https://pypi.org/simple
ffmpeg-python==0.2.0
future==0.18.2
